+++ src/features/settings/components/CreditTabContent.tsx
@@ -1,0 +1,17 @@
+/**
+ * CreditTabContent - 크레딧 사용내역 탭
+ */
+export const CreditTabContent = () => {
+  return (
+    <div className="flex flex-col">
+      <h3 className="font-['Pretendard'] text-[20px] font-semibold text-black">
+        크레딧 사용내역
+      </h3>
+      {/* 구분선 */}
+      <div className="mt-[12px] mb-[40px] h-[1px] bg-[#D1D6DE]" />
+
+      {/* TODO: 크레딧 사용내역 컨텐츠 구현 */}
+      <p className="text-[#6B7280]">크레딧 사용내역 탭 컨텐츠 (구현 예정)</p>
+    </div>
+  );
+};
