+++ src/features/settings/components/SubscriptionTabContent.tsx
@@ -1,0 +1,17 @@
+/**
+ * SubscriptionTabContent - 구독 정보 탭
+ */
+export const SubscriptionTabContent = () => {
+  return (
+    <div className="flex flex-col">
+      <h3 className="font-['Pretendard'] text-[20px] font-semibold text-black">
+        구독 정보
+      </h3>
+      {/* 구분선 */}
+      <div className="mt-[12px] mb-[40px] h-[1px] bg-[#D1D6DE]" />
+
+      {/* TODO: 구독 정보 컨텐츠 구현 */}
+      <p className="text-[#6B7280]">구독 정보 탭 컨텐츠 (구현 예정)</p>
+    </div>
+  );
+};
